backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pymongo==4.10.1
pydantic[email]==2.10.0
python-dotenv==1.0.0
dnspython==2.6.1
